backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
chromadb==0.5.3
sentence-transformers==2.7.0
python-dotenv==1.0.1
pydantic==2.7.4
pymupdf
